requirements.txt
Flask>=2.0.0
Flask-Cors>=3.0.0
python-dotenv>=0.19.0
langchain-community>=0.0.30
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.20
